fix: remove subrouter
