+++ web/src/screens/Settings.tsx
@@ -1,13 +1,6 @@
 import React from 'react'
 import {CogIcon, DownloadIcon, KeyIcon} from '@heroicons/react/outline'
-import {
-    BrowserRouter as Router,
-    NavLink,
-    Route,
-    Switch as RouteSwitch,
-    useLocation,
-    useRouteMatch
-} from "react-router-dom";
+import {NavLink, Route, Switch as RouteSwitch, useLocation, useRouteMatch} from "react-router-dom";
 import IndexerSettings from "./settings/Indexer";
 import IrcSettings from "./settings/Irc";
 import ApplicationSettings from "./settings/Application";
@@ -72,68 +65,48 @@ function SidebarNav({subNavigation, url}: any) {
     )
 }
 
-export function buildPath(...args: string[]): string {
-    const [first] = args;
-    const firstTrimmed = first.trim();
-    const result = args
-        .map((part) => part.trim())
-        .map((part, i) => {
-            if (i === 0) {
-                return part.replace(/[/]*$/g, '');
-            } else {
-                return part.replace(/(^[/]*|[/]*$)/g, '');
-            }
-        })
-        .filter((x) => x.length)
-        .join('/');
-
-    return firstTrimmed === '/' ? `/${result}` : result;
-}
-
 export default function Settings() {
-    let { url } = useRouteMatch();
+    let {url} = useRouteMatch();
 
     return (
-        <Router>
-            <main className="relative -mt-48">
-                <header className="py-10">
-                    <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-                        <h1 className="text-3xl font-bold text-white capitalize">Settings</h1>
-                    </div>
-                </header>
+        <main className="relative -mt-48">
+            <header className="py-10">
+                <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+                    <h1 className="text-3xl font-bold text-white capitalize">Settings</h1>
+                </div>
+            </header>
 
-                <div className="max-w-screen-xl mx-auto pb-6 px-4 sm:px-6 lg:pb-16 lg:px-8">
-                    <div className="bg-white rounded-lg shadow overflow-hidden">
-                        <div className="divide-y divide-gray-200 lg:grid lg:grid-cols-12 lg:divide-y-0 lg:divide-x">
-                            <SidebarNav url={url} subNavigation={subNavigation}/>
+            <div className="max-w-screen-xl mx-auto pb-6 px-4 sm:px-6 lg:pb-16 lg:px-8">
+                <div className="bg-white rounded-lg shadow overflow-hidden">
+                    <div className="divide-y divide-gray-200 lg:grid lg:grid-cols-12 lg:divide-y-0 lg:divide-x">
+                        <SidebarNav url={url} subNavigation={subNavigation}/>
 
-                            <RouteSwitch>
-                                <Route exact path={url}>
-                                    <ApplicationSettings />
-                                </Route>
+                        <RouteSwitch>
+                            <Route exact path={url}>
+                                <ApplicationSettings/>
+                            </Route>
 
-                                <Route path={`${url}/indexers`}>
-                                    <IndexerSettings />
-                                </Route>
+                            <Route path={`${url}/indexers`}>
+                                <IndexerSettings/>
+                            </Route>
 
-                                <Route path={`${url}/irc`}>
-                                    <IrcSettings />
-                                </Route>
+                            <Route path={`${url}/irc`}>
+                                <IrcSettings/>
+                            </Route>
 
-                                <Route path={`${url}/clients`}>
-                                    <DownloadClientSettings />
-                                </Route>
+                            <Route path={`${url}/clients`}>
+                                <DownloadClientSettings/>
+                            </Route>
 
-                                <Route path={`${url}/actions`}>
-                                    <ActionSettings />
-                                </Route>
+                            <Route path={`${url}/actions`}>
+                                <ActionSettings/>
+                            </Route>
 
-                            </RouteSwitch>
-                        </div>
+                        </RouteSwitch>
                     </div>
                 </div>
-            </main>
-        </Router>
+            </div>
+        </main>
     )
 }
